Do not delete channels when deleting all video sources
When doing "Delete All Videosources" in mythtv-setup, mark all channels as deleted instaed of really deleting them.
This was done correct when deleting a single video source but not when deleting all video sources.
See commit 2e26f7f12fce13c155e43ff536b1e09a342e1722 " Preserve data for deleted channels that are still referenced."

+++ mythtv/libs/libmythtv/sourceutil.cpp
@@ -597,8 +597,18 @@ bool SourceUtil::DeleteAllSources(void)
         return false;
     }
 
-    return (query.exec("TRUNCATE TABLE channel") &&
-            query.exec("TRUNCATE TABLE program") &&
+    // Delete all channels
+    query.prepare("UPDATE channel "
+                  "SET deleted = NOW(), mplexid = 0, sourceid = 0 "
+                  "WHERE deleted IS NULL");
+
+    if (!query.exec() || !query.isActive())
+    {
+        MythDB::DBError("Deleting all Channels", query);
+        return false;
+    }
+
+    return (query.exec("TRUNCATE TABLE program") &&
             query.exec("TRUNCATE TABLE videosource") &&
             query.exec("TRUNCATE TABLE credits") &&
             query.exec("TRUNCATE TABLE programrating") &&
